fix: Images appear in search w/o errors now
