+++ imports/plugins/included/product-variant/containers/productGridContainer.js
@@ -2,10 +2,12 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import update from "immutability-helper";
 import _ from "lodash";
-import { Components, registerComponent } from "@reactioncommerce/reaction-components";
+import { compose } from "recompose";
+import { Components, composeWithTracker, registerComponent } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
 import { Session } from "meteor/session";
 import { Reaction } from "/client/api";
+import { Media } from "/imports/plugins/core/files/client";
 import Logger from "/client/modules/logger";
 import { ReactionProduct } from "/lib/api";
 import ProductGrid from "../components/productGrid";
@@ -171,6 +173,47 @@ const wrapComponent = (Comp) => (
   }
 );
 
-registerComponent("ProductGrid", ProductGrid, wrapComponent);
+function composer(props, onData) {
+  // Instantiate an object for use as a map. This object does not inherit prototype or methods from `Object`
+  const productMediaById = Object.create(null);
+  (props.products || []).forEach((product) => {
+    const primaryMedia = Media.findOneLocal({
+      "metadata.productId": product._id,
+      "metadata.toGrid": 1,
+      "metadata.workflow": { $nin: ["archived", "unpublished"] }
+    }, {
+      sort: { "metadata.priority": 1, "uploadedAt": 1 }
+    });
 
-export default wrapComponent(ProductGrid);
+    const variantIds = ReactionProduct.getVariants(product._id).map((variant) => variant._id);
+    let additionalMedia = Media.findLocal({
+      "metadata.productId": product._id,
+      "metadata.variantId": { $in: variantIds },
+      "metadata.workflow": { $nin: ["archived", "unpublished"] }
+    }, {
+      limit: 3,
+      sort: { "metadata.priority": 1, "uploadedAt": 1 }
+    });
+
+    if (additionalMedia.length < 2) additionalMedia = null;
+
+    productMediaById[product._id] = {
+      additionalMedia,
+      primaryMedia
+    };
+  });
+
+  onData(null, {
+    productMediaById
+  });
+}
+
+registerComponent("ProductGrid", ProductGrid, [
+  composeWithTracker(composer),
+  wrapComponent
+]);
+
+export default compose(
+  composeWithTracker(composer),
+  wrapComponent
+)(ProductGrid);
